extract histogram extraction to a new class
